🧠 RAG & Knowledge Graph/core/rag_engine.py
"""
RAG引擎 - 核心处理引擎
"""

import os
import yaml
from typing import List, Dict, Any, Optional
from pathlib import Path
from datetime import datetime


class RAGEngine:
    """
    RAG核心引擎
    
    负责：
    1. 文档管理
    2. 知识检索
    3. 上下文增强
    4. 多源信息整合
    """
    
    def __init__(self, config_path: str = "config.yaml"):
        """初始化RAG引擎"""
        self.config = self._load_config(config_path)
        self.vector_store = None
        self.knowledge_graph = None
        self.preprocessor = None
        self.retriever = None
        
        print("🧠 RAG引擎初始化中...")
        self._initialize_components()
        print("✅ RAG引擎初始化完成")
    
    def _load_config(self, config_path: str) -> Dict:
        """加载配置文件"""
        try:
            with open(config_path, 'r', encoding='utf-8') as f:
                return yaml.safe_load(f)
        except FileNotFoundError:
            print(f"⚠️  配置文件未找到: {config_path}，使用默认配置")
            return self._get_default_config()
    
    def _get_default_config(self) -> Dict:
        """获取默认配置"""
        return {
            "rag": {
                "chunk_size": 512,
                "chunk_overlap": 50,
                "top_k": 5,
                "similarity_threshold": 0.7
            },
            "vector_store": {
                "type": "chroma",
                "persist_directory": "./data/chroma"
            }
        }
    
    def _initialize_components(self):
        """初始化组件"""
        # 这里先创建占位符，后续实现
        print("  - 初始化向量存储...")
        # self.vector_store = VectorStore(self.config)
        
        print("  - 初始化知识图谱...")
        # self.knowledge_graph = KnowledgeGraph(self.config)
        
        print("  - 初始化预处理器...")
        # self.preprocessor = Preprocessor(self.config)
        
        print("  - 初始化检索器...")
        # self.retriever = Retriever(self.config)
    
    def add_document(
        self,
        content: str,
        metadata: Optional[Dict[str, Any]] = None,
        source: str = "manual"
    ) -> Dict[str, Any]:
        """
        添加文档到RAG库
        
        Args:
            content: 文档内容
            metadata: 元数据
            source: 来源（manual, ERP, web, chat等）
        
        Returns:
            添加结果
        """
        print(f"\n📄 添加文档到RAG库")
        print(f"   来源: {source}")
        print(f"   内容长度: {len(content)}字符")
        
        # TODO: 实现完整的文档添加流程
        # 1. 预处理
        # 2. 分块
        # 3. 向量化
        # 4. 存储
        # 5. 知识图谱更新
        
        result = {
            "success": True,
            "doc_id": f"doc_{datetime.now().strftime('%Y%m%d%H%M%S')}",
            "chunks_count": len(content) // self.config['rag']['chunk_size'] + 1,
            "source": source,
            "timestamp": datetime.now().isoformat()
        }
        
        print(f"✅ 文档添加成功 (ID: {result['doc_id']})")
        return result
    
    def query(
        self,
        question: str,
        top_k: Optional[int] = None,
        filters: Optional[Dict] = None,
        use_knowledge_graph: bool = True
    ) -> Dict[str, Any]:
        """
        查询RAG库
        
        Args:
            question: 查询问题
            top_k: 返回结果数量
            filters: 过滤条件
            use_knowledge_graph: 是否使用知识图谱
        
        Returns:
            查询结果
        """
        print(f"\n🔍 查询RAG库")
        print(f"   问题: {question}")
        
        if top_k is None:
            top_k = self.config['rag']['top_k']
        
        # TODO: 实现完整的查询流程
        # 1. 向量检索
        # 2. 知识图谱检索
        # 3. 结果重排序
        # 4. 上下文整合
        
        result = {
            "success": True,
            "question": question,
            "results": [
                {
                    "content": "这是一个示例结果",
                    "score": 0.95,
                    "source": "示例文档",
                    "metadata": {}
                }
            ],
            "total_results": 1,
            "query_time_ms": 100
        }
        
        print(f"✅ 查询完成，找到{result['total_results']}个结果")
        return result
    
    def add_from_file(self, file_path: str, metadata: Optional[Dict] = None) -> Dict:
        """
        从文件添加到RAG库
        
        Args:
            file_path: 文件路径
            metadata: 元数据
        
        Returns:
            处理结果
        """
        print(f"\n📁 从文件添加: {file_path}")
        
        # TODO: 实现文件处理
        # 1. 识别文件类型
        # 2. 提取内容
        # 3. 预处理
        # 4. 添加到RAG
        
        result = {
            "success": True,
            "file_path": file_path,
            "file_type": Path(file_path).suffix,
            "processed": True
        }
        
        print(f"✅ 文件处理完成")
        return result
    
    def get_statistics(self) -> Dict[str, Any]:
        """获取RAG库统计信息"""
        return {
            "total_documents": 0,
            "total_chunks": 0,
            "total_entities": 0,
            "total_relations": 0,
            "storage_size_mb": 0,
            "last_update": datetime.now().isoformat()
        }
    
    def health_check(self) -> Dict[str, Any]:
        """健康检查"""
        return {
            "status": "healthy",
            "vector_store": "ok",
            "knowledge_graph": "ok",
            "timestamp": datetime.now().isoformat()
        }


def test_rag_engine():
    """测试RAG引擎"""
    print("="*70)
    print("  RAG引擎测试")
    print("="*70)
    
    # 初始化
    rag = RAGEngine()
    
    # 添加文档
    rag.add_document(
        content="这是一个测试文档，用于验证RAG引擎功能。",
        metadata={"type": "test"},
        source="test"
    )
    
    # 查询
    result = rag.query("RAG引擎如何工作？")
    print(f"\n查询结果: {result}")
    
    # 统计
    stats = rag.get_statistics()
    print(f"\n统计信息: {stats}")
    
    # 健康检查
    health = rag.health_check()
    print(f"\n健康状态: {health}")
    
    print("\n✅ RAG引擎测试完成！")


if __name__ == "__main__":
    test_rag_engine()








