🧠 RAG & Knowledge Graph/requirements.txt
# RAG & Knowledge Graph - Python依赖
# 版本: v1.0.0

# ========== 核心框架 ==========
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# ========== 向量数据库 ==========
chromadb==0.4.18
# 或使用其他向量数据库：
# faiss-cpu==1.7.4
# qdrant-client==1.7.0

# ========== 文本处理 ==========
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
tiktoken==0.5.2

# ========== 文件处理 ==========
# PDF
pypdf2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8

# Office文档
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
pandas==2.0.3

# 文本文件
chardet==5.2.0
python-magic==0.4.27

# 代码文件
pygments==2.17.2

# ========== 图片处理 ==========
Pillow==10.1.0
pytesseract==0.3.10  # OCR
easyocr==1.7.0       # 多语言OCR

# ========== 音视频处理 ==========
librosa==0.10.1      # 音频分析
SpeechRecognition==3.10.0  # 语音识别
pydub==0.25.1        # 音频处理
opencv-python==4.8.1 # 视频处理
moviepy==1.0.3       # 视频编辑

# ========== 知识图谱 ==========
neo4j==5.14.1        # Neo4j驱动
networkx==3.2.1      # 图分析
pyvis==0.3.2         # 图可视化

# ========== NLP ==========
spacy==3.7.2         # NLP工具
# python -m spacy download zh_core_web_sm
jieba==0.42.1        # 中文分词

# ========== 向量化 ==========
transformers==4.35.2
torch==2.1.1

# ========== 数据库 ==========
sqlalchemy==2.0.23
aiosqlite==0.19.0

# ========== 工具库 ==========
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
httpx==0.25.1
aiofiles==23.2.1

# ========== 数据处理 ==========
numpy==1.24.3
scikit-learn==1.3.2

# ========== 日志 ==========
loguru==0.7.2

# ========== 测试 ==========
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# ========== 其他 ==========
tqdm==4.66.1         # 进度条
python-multipart==0.0.6
beautifulsoup4==4.12.2  # HTML解析
lxml==4.9.3          # XML解析
markdown==3.5.1      # Markdown解析












